requirements.txt
numpy
opencv-python
pypangolin
PyOpenGL